Revert "net/mlx5e: Fix vport admin state to follow port state"
This reverts commit cd7a9f1a29a8dfd1d5181b59df04f7a1b8565295.

In LAG vPort traffic is forwarded to the master vPort and from there it
might be forwarded to the slave nic for performing LAG
This design cause the vPort up/down to be unusable since switching
master vPort to down stops the traffic to both vPorts.

Need to revert this patch in the meanwhile to keep RoCE LAG working.

Issue:1444489
Change-Id: Ibc2b3fa1ffa0fcf5577960064e115b0442693e1d
Signed-off-by: Eran Ben Elisha <eranbe@mellanox.com>