IB/core: Delete lower netdevice default GID entries in bonding scenario
When NETDEV_CHANGEUPPER event occurs, lower device is not yet
established as slave of the master, and when upper device is bond
device, default GID entries are fail to delete.
Due to this, when bond device is fully configured, default GID entries
of bond device cannot be added as default GID entries are occupied by
the lower netdevice. This is incorrect. Default GID entries should
really be of bond netdevice because in all RoCE GIDs (default or IP),
MAC address of the bond device will be used.
It is confusing to have default GID of netdevice which is not really used
for any purpose.

Therefore, as first step, implement
(a) filter function which filters if a CHANGEUPPER event netdevice and
associated upper device is master device or not.
(b) callback function which deletes the default GIDs of lower (event
netdevice).

Signed-off-by: Parav Pandit <parav@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>