RDMA/cma: Protect cma dev list with lock
When AF_IB addresses are used during rdma_resolve_addr(), currently lock
is not held. A cma device can get removed while such list traversal is
in progress which may lead to crash.

Therefore, hold a lock while traversing the list which avoids such
situation.
This patch is not tested. It was discovered during code review.

Cc: <stable@vger.kernel.org> # 3.10
Fixes: f17df3b0d ("RDMA/cma: Add support for AF_IB to rdma_resolve_addr()")
Signed-off-by: Parav Pandit <parav@mellanox.com>
Reviewed-by: Daniel Jurgens <danielj@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>