IB/mlx5: Set uid as part of MKEY commands
Set uid as part of MKEY create command so that the firmware can manage the
MKEY object in a secured way.

The uid for the destroy command is set by mlx5_core.

The fast registration mode (i.e. UMR) can't be used with uid as it
doesn't reach the firmware, as such the regular path is chosen.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>