net/mlx5: E-Switch, Get counters for offloaded flows from callers
There's no real reason for the e-switch logic to manage the creation of
counters for offloaded flows. The API already has the directive for the
caller to denote they want to attach a counter to the created flow.
As such, we go and move the management of flow counters to the mlx5e
tc offload logic. This also lets us remove an inelegant interface where
the FS layer had to provide a way to retrieve a counter from a flow rule.

Signed-off-by: Mark Bloch <markb@mellanox.com>
Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>

 Conflicts:
	drivers/net/ethernet/mellanox/mlx5/core/en_tc.c
	drivers/net/ethernet/mellanox/mlx5/core/eswitch_offloads.c

Change-Id: I8ba84b64b34e0093f7b3ae5c6db363f1ad84ac8c