net: sched: cls_flower: pass tc handle information to hardware
If user specifies skip_sw and the matches of two tc filter commands
are the same, hardware needs to differentiate two cases:
  1) replace
  2) duplicate add command

If the matches are the same and the handles are the same, it is a
replace command. If the matches are the same and the handles are
not the same, it should a duplicate add command.

But now Hardware doesn't have the handle information to differentiate
them. So pass it to hardware.

Issue: 1547947
Signed-off-by: Chris Mi <chrism@mellanox.com>