net/mlx5: Remove HW fte on last dest removal
When the last dest is removed from the fte make sure we also remove
the HW fte. This will prevent a situation where the HW fte isn't really
removed but the user think it is and tries to free the resources being
used by it (like modify header, packet reformat etc).

Issue: None
Change-Id: I0efdec423e12cd8eb9fb6e7c053ac2064433ace7
Signed-off-by: Mark Bloch <markb@mellanox.com>