Revert "PCI: Provide sensible IRQ vector alloc/free routines"
This reverts commit 8a1368f31dcb10966961515f0c5a40e39aa6936b.

In this patch, pci_alloc_irq_vectors and pci_free_irq_vectors are
defined, but not used. Because in order to use them, more patches
need to be backported that is too complex. So we still use old API.
But OFED thinks we are using these APIs and we can't compile OFED
correctly.

Signed-off-by: Chris Mi <chrism@mellanox.com>