net/mlx5e: miniflow, Remove unnecessary rhashtable_remove_fast() call
We will remove the entry from the hash-table once mlx5e_del_miniflow()
is called. Please note this happens after the flow is removed from the
hardware.

This will prevent a scenario in which we might try to merge and offload
the same connection before it is removed from hardware.

Change-Id: I2f66b61d0d96c16120a4d4f638ab90dc9e1a7164